+++ src/common.h
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#define CORNAMUSA_VERSION "0.1.0"
+#define CORNAMUSA_VERSION "0.4.0"
 #define CORNAMUSA_VERSION_MAJOR 0
-#define CORNAMUSA_VERSION_MINOR 1
+#define CORNAMUSA_VERSION_MINOR 4
 #define CORNAMUSA_VERSION_PATCH 0
 
 #endif /* CORNAMUSA_COMMON_H */